Move showSignedOutNotification() / showSignedOutAlert() to SessionNotificationHelper
